Change `ir` format to emit earlier.
The overall plan is to move all the tranforms into the tint backends.
This means, the `tint` command will end up not running any transforms
itself. Once that is done, then emitting the `ir` from the `--dump-ir`
flag would be the same as emitting the `--format ir`. So, pre-emptively
move the format code up with `--dump-ir`.

Change-Id: Ibd00d34f1d046962bd9a7e30330a16ff934e5624
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/216594
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>