[tint][ir][fuzz] Refactor IR fuzzer infra to impl pre/post validation
This change hoists the bespoke code from a variety of fuzzer
implementation for doing post-run validation up into the fuzzer runner
code, so that copy-paste errors and capability omissions are more
difficult.

Fuzzer creation now takes in either a single Capability set that is
used for validation before and after a run, or a pair of Capability
sets, one for before validation and one for post validation.

This required having the fuzzer impl function return whether or not
the run was successful, thus whether or not post-validation should
occur. Instead of using a simple bool, the Result<SuccessType> that
many transform impls return was instead used, so these impls could be
passed into the fuzzing infra with less boiler plate.

Some fuzzing passes that previously did not perform a post-validation
step now do with this change.

Fixes: 380282495
Change-Id: I0372a326384d8f29aa4a61dd02b49a61dc2efc7d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/216534
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Reviewed-by: James Price <jrprice@google.com>