[msl] Skip TINT_ISOLATE_UB for finite loops
Use the loop analysis helper to detect loops that are guaranteed to be
finite, and do not use this macro for them since we know they are
safe.

Suppress some new CTS failures which appear to be due to an AMD driver
bug.

Bug: 379009123
Change-Id: I5a541b4c81ac2ffc3842038fa585b9404d8574d5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/216617
Auto-Submit: James Price <jrprice@google.com>
Commit-Queue: David Neto <dneto@google.com>
Reviewed-by: David Neto <dneto@google.com>