[dawn] Begin deprecation of SubgroupsF16 feature; step 1
This change drops Subgroups support for Vulkan devices that support
ShaderF16 but not SubgroupsF16. This is a very narrow set of devices
estimated to be 4%. [0]

No immediate changes are required for D3D or Metal as support for
extended types (f16) is consistent with subgroup operations.

[0]
https://github.com/gpuweb/gpuweb/blob/main/proposals/subgroups.md#enable-extension

Bug:380244620
Change-Id: I81c31fa8b3ce5d664e98f3b7a6c74c8d512c3ed6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/217474
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>