[tint] Begin deprecation of 'subgroups_f16' enable. Step 2
Step 1:
https://dawn-review.googlesource.com/c/dawn/+/217474

In this step 2 of deprecation we:
Allow for f16 subgroup operations with only 'f16' and 'subgroup'

NOTE: CTs tests these requirements so we need to put in expectations
to silence the 'subgroups_f16' requirement until the CTS change rolls.

This is actually the most important change. The next important change
is to fix the CTS. After this we can fully deprecate and remove the
'subgroups_f16' at our leisure.

https://github.com/gpuweb/gpuweb/blob/main/proposals/subgroups.md#enable-extension

Bug:380244620
Change-Id: Ia24054693b420dc436867f0aded866c2657e521a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/217294
Commit-Queue: Peter McNeeley <petermcneeley@google.com>
Reviewed-by: James Price <jrprice@google.com>