Add dawn-texel-copy-buffer-row-alignment feature
The current alignment is always 256. This more likely causes the padding
gaps between the rows in the buffer, and incurs an extra copy in the
BlitTextureToBuffer helper. With the new feature this CL introduces, the
alignment is loosened to 4 bytes on D3D11 backend.

Bug: chromium:378361783
Change-Id: I7484a45b283f69021d14471edaa2a7d3d4dfd9ee
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/214478
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Jie A Chen <jie.a.chen@intel.com>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>