[tint][ir][fuzz] Add missing capability for GLSL fuzzing pass
The `glsl.TexturePolyfill` transform converts 1D handles to 2D ones
and stripes off the binding information, which requires
kAllowHandleVarsWithoutBindings to pass validation.

This is expected behaviour, since this capability is added in the
unittests, so adding it to the fuzzer capabilities, so that the
fuzzing pass doesn't spuriously fail.

Issue: 340578705
Change-Id: I13ace1f05ec38fe2b32e970be6b7ddecfedee6c7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/217954
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: James Price <jrprice@google.com>