tint: fix memory leak in Command::LookPath on Windows
`Command::LookPath` calls `ExecutableExists` which uses
`MapViewOfFileEx` to check if the file is executable or not. But we were
not calling `UnmapViewOfFile` on the returned address, so this would
leak. LookPath in mainly called when looking for dxcompiler.dll, which
we only do from the tint executable and the fuzzers. As the tint exe
only processes one input at a time, we didn't notice this leak; but the
fuzzers process multiple inputs, so the leak became obvious with the
fuzzer crashing when OOM.

Bug: 381549729
Change-Id: I99f8c75844a195458c91d49928eee41a15ea0307
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/218137
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>