[tint] Always perform requested renames in Renamer
The Renamer was not applying `requested_names` when only renaming
keywords. Fix the `should_rename` check to consider `requested_names`.

This is required in order for Dawn to work with symbol naming disabled
when it is using `requested_names` to set the entry point name.

Bug: 375192379
Change-Id: I188608872dfedfda02a807137c694289e1e78a04
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/218138
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>