[tint][ir][val] Restrict types in root allowed by kAllowOverrides
Restricts the instruction types that are allowed in the root block
when kAllowOverrides is set. This doesn't do the walk up to make sure
that the specific values/calls are allowed. This supersedes a number
of previous patches that were playing whack-a-mole.

Fixes: 382453612
Fixes: 382422295
Change-Id: Iccd2145afb29bf61b2dc7ccbfd97b08648e39208
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/218354
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>