Remove AST transforms from WGSL generation.
This CL removes the use of the AST transforms in the WGSL generation.
The transforms will eventually be moved into the respective backends as
IR transforms, the WGSL backend does not use the IR as it's just dumping
out the AST. So, remove the transform usage.

Bug: 380043961
Change-Id: Ifc8f78eda4c114861692c585c9add2ca434716d7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/218295
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>