[gl] Rename buffer variables in the printer
This fixes an issue where user identifiers were still be included in
declarations that we emit in the backend, which is potential security
issue.

This also prepares the backend for handling symbol renaming entirely
inside the printer, instead of with a transform run by Dawn.

Fixed: 382744100
Change-Id: Ieacb9a3b5c5ed65bda0af8b4c77e0b95b3c3edd0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/218617
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: James Price <jrprice@google.com>