[msl] Move symbol renaming into the backend
Instead of running the AST renamer transform in Dawn's Metal backend,
we just modify the names when emitting them in the printer.

This is a little cleaner than using a transform, as transforms cannot
rename structures and struct members without requiring
const_cast. This also keeps the list of MSL keywords local to the
printer, and means we can retain the meaningful names throughout the
`raise` process.

Bug: 380043958
Change-Id: Ib5ae9b36947826928fd7babaed9892d06a8c0c29
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/219295
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: James Price <jrprice@google.com>