[glsl] Move combined texture sampler naming to printer
Instead of setting the name for combined texture sampler variables in
the TexturePolyfill transform, apply them in the Printer instead. We
do this by adding a new instruction for combined texture sampler
variables that the polyfill transform generates. The printer can then
perform the map lookup to find the desired name.

This removes the need to preserve the names that are held in the IR
module, which in turn simplifies the renaming process.

Bug: 380043958
Change-Id: I47b50927847eaaf93384049457cf3d01da48e789
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/219234
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: James Price <jrprice@google.com>