[gl] Move symbol renaming into the backend
Instead of running the AST renamer transform in Dawn's OpenGL backend,
we just modify the names when emitting them in the printer.

This is a little cleaner than using a transform, as transforms cannot
rename structures and struct members without requiring
const_cast. This also keeps the list of GLSL keywords local to the
printer, and means we can retain the meaningful names throughout the
`raise` process.

Bug: 380043958
Fixed: 42241122
Change-Id: If388c9456d413899f8546eb24cff7819e7caf146
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/218619
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: James Price <jrprice@google.com>