[hlsl] Move symbol renaming into the backend
Instead of running the AST renamer transform in Dawn's D3D backends,
we just modify the names when emitting them in the printer.

This is a little cleaner than using a transform, as transforms cannot
rename structures and struct members without requiring
const_cast. This also keeps the list of HLSL keywords local to the
printer, and means we can retain the meaningful names throughout the
`raise` process.

Bug: 380043958
Change-Id: Ia654c14eabefe834a438bce841f83d7805602e9a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/219296
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: James Price <jrprice@google.com>