[tint] Add a const param validator
Override parameters need to be const evaluated for pipeline errors
after substitute overrides is called. We intend to call this from the
IR so this mean we need to validate parameters in the IR.

We may still need to do a follow up for Access/LVE/SVE too
(for OOB indices on non-constant arrays/vectors/matrices).

Bug: 382300469
Change-Id: I8fe1272e11f991f7047a327ce8692b9826604f3c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/219734
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>