Fix crash when calling `GetBindGroupLayout()` on null bindGroupLayout
This patch fixes a crash issue when calling `GetBindGroupLayout()` on a
null bind group layout item in the pipeline layout used to create the
pipeline. According to the latest WebGPU SPEC, the parameter `index` is
valid as long as `index < the size of this.[[layout]].[[bindGroupLayouts]]`,
so calling `GetBindGroupLayout()` on a null bind group layout item in the
pipeline layout should be legal and return an empty bind group layout.

Bug: 377836524, 42241530
Test: dawn_unittests
Change-Id: I57bc9a136e3da01e8442f6752c98bfa09eef6f2f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/220455
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>