Update `InsertAfter` naming.
The name `InsertAfter` is confusing as it can receive a `FunctionParam`
or `BlockParam` as argument, but it inserts into the _block_ after those
instructions, not the instruction itself.

Update the name to `InsertInBlockAfter` to make it clearer where the
instructions will be placed.

Bug: 361142936
Change-Id: I5ae9b180e274bd6c0589d698a678571de7384347
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/220375
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>