Validate UTF8 conversion is in range.
When converting UTF8 characters, we'd check the various bytes but we did
not validate that the resulting value was within range of the codepoints
supported by the number of bytes decoded.

This CL adds a check to the UTF-8 decoder to validate that the resulting
`uint32_t` value is within range for the number of source bytes.

Bug: 388039766
Change-Id: Ie7ed6a19e84f04356e3fdeba578281cf966145f6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/221375
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>