Mark some enum entries as `@internal`
The defs file allow marking things as `@internal` which removes the
parsing code and does not add them to the list of suggested spellings.

This CL updates the `in` and `out` address spaces along with the
`point_size` builtin to be `@internal`. A series of tests are removed
which are from either old SPIR-V transform code, or ways of writing WGSL
which we no longer generate.

Bug: 42250952
Change-Id: I2aba36123bf0745652d90be4fa336137ebdda420
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/221436
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>