Add --enforce-default-limits to run-cts command
It's important to test CTS tests with --compat and --enforce-default-limits
because some devices have 0 storage buffers and 0 storage textures in
vertex and fragment stages and it's easy to make a test that will fail
if those are 0.

The CTS will force them to the default with this flag so adding an
argument so the flag can bu used from run-cts.

Bug: 389099880
Change-Id: Ie8b4e6b979a4540ba6684e2be8cc0c76d80e3516
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/221754
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Gregg Tavares <gman@chromium.org>