[spirv-reader][ir] Add support for UniformConstant address space.
Add conversion of the `UniformConstant` address space in SPIR-V into
the `handle` address space in IR. Preliminary support for the spirv
`Sampler` type is added as well in order to support the test case. The
sampler code will need to be extended to also support comparison
samplers.

Bug: 42250952
Change-Id: Ib56d58215a42cf8f4b67f237cee1adb84b680328
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/220494
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>