[tint][spirv] Check module + options before generating
Move the `CanRun()` function from the writer fuzzer into the writer as
a helper function. This can then be used by both the fuzzer and the
Tint executable to check whether the backend can actually handle the
module+options.

This yields much more user friendly error messages from the Tint
executable when trying to generate backend code for a module that
contains features that are not supported by that backend
(e.g. pixel_local variables with the SPIR-V backend).

Bug: 376572262
Change-Id: Icf2b73f2144d53ea66e7943e2341387d28616ec3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222114
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>