[tint][hlsl] Check module + options before generating
Move the `CanRun()` function from the writer fuzzer into the writer as
a helper function. This can then be used by both the fuzzer and the
Tint executable to check whether the backend can actually handle the
module+options.

This yields much more user friendly error messages from the Tint
executable when trying to generate backend code for a module that
contains features that are not supported by that backend.

Bug: 376572262
Change-Id: I23d1cea04ab2a3cde769f577a4287f7649f535cf
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222135
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>