[tint][ir][val] Check that conversion for specific types exists
This CL introduces looking up conversion entries in the intrinsic table
for the convert call. It does not enforce that entry is a converter (not
a constructor), so is less restrictive then the intended final state.

A follow up CL will add the 'is converter' check and update various
tests.

Bug: 389589696
Change-Id: I2726e6546545fa5cb719343f04e3c286290b727a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222115
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>