dawn.node: add iterators for features
In the browser,

```js
...  = [...adapter.features]
```

works but not in dawn.node. Various tests now require this to
work so add it.

Note that there are still other interop issues.

In the browser this works

```js
   device = adapter.requestDevice({requiredLimits: adapter.limits});
```

This throws in dawn.node that the value of requiredLimits is not an array.

[test](https://gist.github.com/greggman/b2ecc05affb560dc88e0c81263112ee6)

Change-Id: I9e9fd14a75372cbe713335a82e948ea0d027afa6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222274
Commit-Queue: Gregg Tavares <gman@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>