GL: Fall back to TexImage texture allocation for some formats
GL_BGRA8_EXT is an edge case format that is not supported for
glTexStorage when GL_EXT_texture_storage is not present, even after
glTexStorage became core in ES 3.0.

Add a new capability flag to GLFormat and a fallback path in
AllocateTexture which initializes the levels individually with
glTexImage.

Bug: 346356622
Change-Id: If5a147f67d3fdb756cff4cdb6effe02551c2e316
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/221634
Reviewed-by: Stephen White <senorblanco@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>