webgpu.h: Add "emscripten" prefix to surface source HTML selector
While adding the "emscripten" prefix to the Emscripten structure and
sType for getting a surface from a canvas HTML selector, it also makes
sure structures and enums start with "emsscripten" if they're tagged as
so, and don't if they're not tagged. A similar approach will be used for
"dawn" tagged ones in the future.

Bug: 42241174
Change-Id: Ib29948a220fb6cbbcbf114d0222a48a6d4c056fd
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222654
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Fr <beaufort.francois@gmail.com>