[tools] Add MemMapOSWrapper
Adds MemMapOSWrapper and its related interfaces. This is the standard
implementation meant for use in testing, although tests can still create
test-specific implementations if necessary.

This is adds a dependency on Afero, which seems to be the go-to project
for mocking out the filesystem in Go. Since Go's built-in os.File is a
struct instead of an interface, all OSWrappers end up needing to use
afero.File instead of *os.File for file handles. This should be a
drop-in replacement for all common uses, but does mean that some less
common cases are not supported, such as testing with symlinks.

Bug: 344014313
Change-Id: I8bd040397db2b4bffeec16f6f47e92d3014f3d9f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/222914
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@google.com>
Auto-Submit: Brian Sheedy <bsheedy@google.com>