metal: Use default storage mode for IOSurface textures
The best practice is to use the default storage mode which is "shared"
on Apple GPUs and "managed" on non-Apple GPUs. Using a non-default
storage mode could have implications on how IOAccelerator allocations
are made for the IOSurface e.g. managed implies lack of coherence
between CPU and GPU and needs to be explicitly synchronized - it's not
clear if this can cause an extra allocation on Apple GPUs.

Bug: 388872587
Change-Id: Ief7905593d08251b8f4626e5f14b06e6e98a3c6a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/223275
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>