[inspector] Fix issue with finding texture,sampler pairs.
When looking for the entry points associated to samplers and textures
we'd find the calls which used a sampler, and then add any entry point
which was reachable up the call chain from that usage.

This is incorrect in the face of a forked call pattern:

```
            s
          /  \
         /    \
       /        \
 s(t0, s0)    s(t1, s0)
     /            \
  vert           frag
```

In this case, we'd associate both `t0,s0` and `t1,s0` to both the vertex
and fragment shaders. Instead, we should only associate `t0,s0` to the
vertex shader and `t1,s0` to the fragment shader.

This Cl changes the routine such that instead of getting the entry
points early it waits until we've found both the texture and sampler in
the call stack and uses that function as the basis for determining the
entry points.

Bug: 380433758
Change-Id: Ibe313404417ba77a0ed5467c4d1b20f188fe47a6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/223534
Reviewed-by: David Neto <dneto@google.com>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>