Test for TEXTURE_BINDING vs STORAGE_BINDING issue
The GL backend has an issue if you used a texture as a TEXTURE_BINDING
with a limited subview and then use the same texture has a
STORAGE_BINDING with a different subview. In the backend the first
usage sets TEXTURE_BASE_LEVEL and TEXTURE_MAX_LEVEL on the texture
and this affected the 2nd usage.

This is just a test for this issue. It's suppressed because it fails
on various backends.

Bug: 392121637,392121643,392121648
Change-Id: I2d048ee3d6a065ffb1e1f397f439c335ce96f4b7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/223594
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Auto-Submit: Gregg Tavares <gman@chromium.org>
Commit-Queue: Gregg Tavares <gman@chromium.org>