[tint] Rename array to runtime_array in intrinsics
We need to add fixed-sized arrays to the intrinsics file for some of
the new subgroup_matrix builtins, but the intrinsics parser does not
support having two different template parameterizations of the same
type.

Rename `array` to `runtime_array` and then add logic to emit it as
`array` when generating tests.

Bug: 348702031
Change-Id: I78948c3576935d726f2580b34bd22d9130446a46
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/224054
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>