[tools][cts] Add credentials-check
Adds the credentials-check subcommand to tools/run cts, which checks
that the current user has the required permissions for various remote
services. This check is also run at the beginning of a roll attempt to
help avoid issues where a roll runs for several hours before failing due
to permission issues.

There are several services which are explicitly not checked at this time
due to the APIs not having a good way to check user permissions.

Bug: 349798588
Change-Id: Iee3d270f3e86ec25cd7cf300244997b0dfae4258
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/223799
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@google.com>