metal: Specify storage binding usage for IOSurface textures explicitly
Add a allowStorageBinding flag on SharedTextureMemoryIOSurfaceDescriptor
which specifies whether TextureUsage::StorageBinding is allowed and if
MTLTextureUsageShaderWrite is set on the IOSurface wrapping MTLTexture.
The flag defaults to true to match current behavior, but we will change
it to false once Chromium migrates to using the flag. Chromium will set
the flag based on whether SHARED_IMAGE_USAGE_WEBGPU_STORAGE_TEXTURE
usage is set on the IOSurfaceImageBacking.

Bug: 388872587
Change-Id: I8cd78711cf8600b77a3d14b58eaa6d026afe83d1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/223216
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>