[msl] Handle subgroupMatrixLoad builtins
Replace them with simdgroup_load intrinsics in the builtin polyfill
transform. The MSL intrinsics take a pointer to the first element, and
have a `matrix_origin` parameter that must be a `ulong2`. We need to
declare a temporary variable to load the matrix into.

Bug: 348702031
Change-Id: If49f83f69d82dd9104c69ad482e738113f6f71c2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/224255
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>