[spirv-reader][ir] Fix `ModfStruct` results.
When converting the `ModfStruct` instruction we swapped the result type
for the internal `modf` result type. This works to update the
instruction, but later fails if there is an instruction which depends
on the original spirv type (i.e. creating an `OpVariable` and storing
the result of the `ModfStruct`.

This CL updates the parser to handle the re-creation of the original
SPIR-V type and treats that construct as the result of the original
SPIR-V `ModfStruct` instruction.

Bug: 42250065
Change-Id: Ie066a1c7fc6a86d3e69981c6937cd0e6d98421c7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/224754
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>