[spirv-reader][ir] Add support for GLSL std450 FrexpStruct
Add a conversion of the FrexpStruct GLSL method to the WGSL frexp
equivalent. The spirv version allows signed or unsigned as part of the
result struct but WGSL only allows signed. Convert as needed. This is
done directly in the printer as we need to access the original SPIR-V
structure type which we'd lose otherwise.

Bug: 391673570
Change-Id: Ic9a571ed17d3a198ee052552c8f5346d43ed4665
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/224794
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>