[spirv-reader][ir] Support `OpKill`
Translate an `OpKill` SPIR-V instruction into a `discard; ret` IR
instructions. The semantics of `OpKill` don't currently exist in WGSL,
the `discard` instruction is more like `OpDemoteToHelperInvocation`. The
`discard; ret` is similar, although not completely the same, but it's as
close as we can currently get.

Bug: 391487196
Change-Id: I8abd5526b866e74f90a3c5cba06cf2001837ab28
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/225094
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>