[tint] Wrap accessing argc & argv to remove suppressions
This adds in an utility for converting argc & argv passed into main()
into a tint::Vector of the arguments.

This is done because it is a known issue that accessing the c-style
argv will always cause `-Wunsafe-buffer-usage` to fire, and this
wrapper isolates this known needed suppression into a single location
instead of being scattered across the code base.

This CL cleans up a bunch of main.cc files using this new utility,
though there is a couple exceptions. In remote_compile/main.cc the
suppression is just narrowed, since there is non-main related issues
in that file. There are some other main.cc files, like for the
benchmark tool, that use a different broader suppression that are not
touched.

Bug: 394825124
Change-Id: I92ea0fbc0012ef0900214a1eaefc9fb89a63553d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/224915
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>