Emit WGSL errors as part of GPUValidationError
This fixes the issue that if you want to wrap the error handling
for things like WESL you can not stop Chrome/Dawn from emitting
messages that are not relevant to your environment.

Without this change, you get a message that says something that is
effectively like "error at line 123 ..." with no way to stop the error
from making it to the user. If you're using some kind of preprocessor
then "line 123" might make no sense and ends up being confusing to
downstream users.

With this change, it's possible to wrap the API and capture the
error and then synthesize your own error with the correct line
and other info, for example if you wanted to polyfill source map
support.

Note that, given that the message is moved to the GPUValidationError
mechanism and given those make it to the console by default, no
info is lost. Users still get useful error messages if they do
nothing.

Bug: 40286992
Change-Id: Ic22d6d21731ac59706e06d517657cf98f0912e3e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/226214
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Gregg Tavares <gman@chromium.org>