[dawn][native] Better control lifetime of transient uploads.
Instead of requesting an upload region that's returned from the
DynamicUploader, use a pattern where a functor is passed to make the
DynamicUploader tightly control the lifetime of it.

In follow-up CLs it will be used to determine how much data is pending a
submit and trigger opportunistic flushes.

Bug: 42242066
Change-Id: I860f74e79834217341523af237312eadb37bc7c4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/226415
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Jiawei Shao <jiawei.shao@intel.com>