Change check to assert for VVL error
On WebView we are receiving VVL errors from the driver when Vulkan
object labels are disabled. This results in a DAWN_CHECK failing and
WebView crashing. It's not entirely clear why the driver is sending
messages with VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT set, as VVL
should be disabled in release builds, but there is some discussion in
the bug.

Either way the error handler is just to alert developers that something
is wrong. Change it back to DAWN_ASSERT() to avoid crashing in release
builds.

Bug: 397098430
Change-Id: I9a9afc76a2a27de4641ed658a250518c5d4aa52c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/226678
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>