Add a dawn wire client API to request a device's id and generation
Expose an API in dawn::wire::WireClient to request a device's id and
generation, returned through the dawn::wire::Handle struct.

This change is required for http://crrev.com/c/5833743. The device id
and device generation are passed from Blink to the WebNN service so
that the WebNN service can lookup the native Dawn WebGPU device.

Change-Id: Ia98d6f9e9f4152f47402a6717fa9cc8412a17d1f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/205174
Commit-Queue: Patrick To <patrto@microsoft.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>