[dawn][metal] Increment the submitted serial only after the commit
The logic was slightly unclear previously because the serial was
incremented before the submit. No other operation used the signal but if
some did, they could assume they used the next pending serial instead of
the one that would be used for the commit.

This should be a noop change.

Bug: 42242066
Change-Id: I1904ac8f41a8a5530ea72f2eb14c15939d94538c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/227054
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Auto-Submit: Corentin Wallez <cwallez@chromium.org>