Skip building mCacheKey when not needed
RenderPipeline[Vk]::mCacheKey contains all shader SPIR-V and all the
Vulkan creation struct data. This results in a large mCacheKey size with
graphite, it's observed to be in tens of kilobytes and it quickly adds
up to megabytes for all of the RenderPipeline objects that are kept
alive.

mCacheKey is only used for serializing VkPipelineCache objects into
BlobCache. This path isn't used with the monolithic pipeline cache so
it's all unnecessary memory usage. Only need to build the cache key if
Toggle::VulkanMonolithicPipelineCache isn't enabled.

Bug: 397934765
Change-Id: I99c2a0ed819a4fe364e71aac7ef75fb470cd6102
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/227154
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>