Tint: fix nested negation unary exp parsing, better prefix increment error
This CL split kMinusMinus for unary expression so that nested negation
expression like (a----1) can be correctly parsed as (a - -(-(-(1)))).
This CL also make error message more clear for misusing prefix increment
in variable updating statements, and add more variable updating
statement unittest including `a++`, `a--`, `++a`, `--a`, `_ = ++a`,
`_ = --a`, `_ = a++`, and `_ = a--`.

Bug: 380168990
Change-Id: I98a579ece15a0696b350b82023e7f574d53b40c3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/227034
Commit-Queue: David Neto <dneto@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: David Neto <dneto@google.com>