Vulkan: Fix subgroup matrix test case failure on Intel Win Vulkan
From the design doc(https://docs.google.com/document/d/1mwntmsm2nM5cAKmCVqFWMASdzvlDZR-G2LwbHKfjWjs/), the behavior is undefined if the stride of subgroupMatrixLoad or subgroupMatrixStore is less than the number of
rows when col_major is true. In the loading-related code, col_major is set to false, but the stride is set to M (number of rows) for the left-hand side (lhs) and K (number of rows) for the right-hand side (rhs). Similarly, in the storing-related code, col_major is set to false, but the stride is set to M (number of rows). This change ensures that col_major is set in accordance with the design document.

Bug: 396741026
Test: dawn_end2end_tests
Change-Id: Icc80a747fd62d05bed5f8e23244e9caafd18bb6d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/227694
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: Jiawei Shao <jiawei.shao@intel.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>