D3D12: Fix warning of invalid small resources alignment
If we build Dawn end2end tests on SDK 10.0.22621.0 and run it on
another device with SDK 10.0.26100.0, the warning of
CREATERESOURCE_INVALIDALIGNMENT_SMALLRESOURCE will not
be filtered because this message id is not available on 22621 and was
not added to the deny list during the compilation. Redefined a new
message id using the value on 22621 for workaround.

Bug: 381742470

Change-Id: Ib4017e76b19b274ca7cd0e3152b7415f8750881f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/228234
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Jiawei Shao <jiawei.shao@intel.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>