[spirv] Only add explicit layout decorations on types that need them
Find all of the array and structure types that require explicit layout
decorations before emitting any instructions, so that we can add the
layout decorations when the types are emitted.

This does not address the case where a type is used in multiple
address space, which will be handled by a new transform that forks
those types.

Bug: 42252012
Change-Id: I4ee37d36d97c729b73fef89782881e1557fbb0ce
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/228834
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>