tint: Improve error message for module-scope function calls
The previous error message for calling a function at module scope
was misleading, stating "functions cannot be called at module-scope".
This is inaccurate as built-in constant functions are allowed.

This CL updates the error message to be more specific and helpful,
indicating that "Custom functions cannot be called at module scope.
Only built-in functions are allowed here."

Bug: tint:42250870
Change-Id: I31fcb15ac9a214f413a5edf75264aee04c9094b6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/228754
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>