[tint] Minor fix to avoid checking this for null
The 'As' member function inadvertently was calling a free function
that checked this against null. This produced in-determinant behavior
as some compilers could remove this redundant null check.

With this change we should always dereference 'this' which should
produces a segfault in the cases of this being null. (Although also
technically dereference of null is UB)

Change-Id: I94c49a660db916aeaa11533c567a00efa32bd6ba
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/229614
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>