[spirv] Handle arrays in ForkExplicitLayoutTypes transform
Add a new `ExplicitLayoutArray` type for the SPIR-V backend to use to
identify arrays that need to be emitted with layout decorations. This
allows us to define another array type that is distinct from the
equivalent array that will not have the decorations.

Update the transform to create these new array types, and to convert
them to/from the original array types.

Bug: 42252012
Change-Id: I92cd7179472df6b9948fd9d809e1ad601cf64191
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/229977
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>