[spirv] Add struct flag for explicit layout decorations
Instead of running a pass in the printer to determine which structs
will need explicit layout decorations, just have the
`ForkExplicitLayoutTypes` transform add a struct flag to signal that
information to the printer, since it already knows which structs will
need the decorations.

This currently requires a `const_cast` due to long-standing tech debt
that causes types to be immutable in IR transforms.

Bug: 42252012
Change-Id: Ifb0aa90a47fb92ec30559f0d76c9b92adc959e7c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/230134
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>