[tint][inspector] Walk globals only once in GetResourceBindings
Previously many bespoke functions walked the globals computed a list of
ResourceBinding, then results were aggregated. Instead walks globals
once and for each of them decide if/how it should be turned into a
ResourceBinding.

This will simplify the future addition of support for BindingArray that
will only need to be added in one place.

end2end tests for MSL are updated because the order of bindings is
modified by the sort in GetResourceBindings (that's used by
FlattenBindings)

Bug: 393558555
Change-Id: I795d4a60c444415bf5f69a513cbe4d318363bd42
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/229595
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>