[spirv-reader][ir] Make IR OverrideIds optional.
With the addition of `OpSpecConstant*` instructions there is a need to
allow an `Override` in the IR to not have an ID associated. This means
that the override could not be substituted in the given IR, but it could
not be substituted in SPIR-V anyway. The WGSL path allows the override
to be set using the name instead of the ID.

Bug: 376304012
Change-Id: I9175faed76f0300f85ee428b83aa259fe85522b8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/230334
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>