Make sure objects are deleted when idle
The Vulkan backend enqueues objects for deletion in FencedDeleter after
the pending command serial is completed. This works until idle and no
more work is submitted and the submitted/completed serial is never
incremented.

Have ReduceMemoryUsage() check if there any Vulkan objects pending
deletion. If there are then it will potentially submit and then tick.
If there was a vkQueueSubmit() the last submitted serial won't become
completed immediately, all work must finish on the GPU first, and
objects may still be pending deletion. Add a bool to ReduceMemoryUsage()
to signal to Chrome that there is more deletion work to be done in the
future and ReduceMemoryUsage() needs to be called again after a short
delay.

Bug: 398193014
Change-Id: I0207522ac501cbba18479f27a4113bd39598855b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/231174
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>