[glsl] Only apply BGRA swizzle to `in` address space.
Currently the `bgra_swizzle` values are getting applied to all address
spaces. This means if we have a `location(0)` in the `in` address space
and a `location(0)` in the `out` address space we'll apply the BGRA
swizzle to both of them. Is then can cause issues if the output address
space item isn't a `vec4f` as the assignment will fail.

Change the `bgra_swizzle` code to only apply to `in` address space
locations.

Bug: 389559096
Change-Id: I992ff2612627644c791913a65d9c6e25a9aa8a34
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/231494
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>